las-viewer/requirements.txt
fastapi
uvicorn[standard]
asyncpg
databricks-sdk>=0.81.0
numpy
scipy
